Break off a big chunk into a lib
